requirements.txt
colorama==0.4.6
iniconfig==2.0.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pytest==8.3.3
tqdm==4.67.1
